requirements/base.txt
celery==4.0.2

Django==1.11.1
bugsnag==3.1.1
django-filter==1.0.2
django-rest-swagger==2.1.2
django-push-notifications==1.5.0
djangorestframework==3.6.3

whitenoise==3.3.0
dj-database-url==0.4.2
dj-email-url==0.0.10

PyMySQL==0.7.11
PyYAML==3.12
pytz==2017.2

django-redis==4.8.0
redis==2.10.5

django-fsm==2.5.0
django-fsm-admin==1.2.4
django-fsm-log==1.3.0
python-i18n==0.3.1
sqlparse==0.2.3
arrow==0.10.0
pillow==4.2.1
xlwt==1.2.0
django-health-check==2.3.0
django-two-factor-auth==1.6.2
django-mathfilters==0.4.0
newrelic==2.88.1.73
boto3==1.4.4
intervaltree==2.0.4
pyexcel-io==0.5.1
pyexcel-xlsx==0.5.0.1

django-celery-beat==1.0.1
jinjasql==0.1.6
django-storages==1.6.3
Logentries==0.17
natsort==5.1.0
jsonfield==2.0.2
django-ajax-selects==1.6.1
